requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Flask-Mail==0.9.1
email-validator==2.0.0
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0

# PostgreSQL support
psycopg2-binary==2.9.7

# Additional production dependencies
redis==4.6.0
blinker==1.6.3
